+++ package.json
@@ -90,13 +90,17 @@
     "gulp": "^4.0.0",
     "html-webpack-plugin": "^5.1.0",
     "husky": "^1.3.1",
-    "jasmine": "^3.0.0",
+    "karma": "^6.3.12",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-jasmine": "^4.0.1",
+    "karma-webpack": "^5.0.0",
     "minimist": "^1.2.5",
     "outline-i18n": "Jigsaw-Code/outline-i18n#v0.0.7",
     "postcss": "^7.0.29",
     "postcss-rtl": "^1.7.3",
     "prettier": "^1.19.1",
     "pretty-quick": "^2.0.1",
+    "puppeteer": "^13.1.2",
     "simple-plist": "^1.1.0",
     "style-loader": "^2.0.0",
     "ts-loader": "^8.0.14",
