+++ src/www/app/app.spec.ts
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {unwrapInvite} from './unwrap_invite';
+import {unwrapInvite} from './app';
 
 describe('unwrapInvite', () => {
   it('ignores empty string', () => {
